% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Augmentation.R
\name{AUC.Area.Aug}
\alias{AUC.Area.Aug}
\title{AUC Calculation for Augmentation Estimator.}
\usage{
AUC.Area.Aug(time, status, idx, subj_covars, mu, tau, calc_var = TRUE)
}
\arguments{
\item{time}{Observation time.}

\item{status}{Status, coded as 0 for censoring, 1 for event, 2 for death.
Note that subjects who are neither censored nor die are assumed to remain
at risk throughout the observation period.}

\item{idx}{Unique subject index.}

\item{subj_covars}{Per-subject covariate data. Should include `idx` and
covariates only.}

\item{mu}{Grand mean.}

\item{tau}{Truncation time.}

\item{calc_var}{Calculate analytical variance?}
}
\value{
List containing:
\itemize{
  \item Tabulated mean cumulative function 'mcf'.
  \item Estimated 'area' and its variance 'var_area'.
  \item 'sigma', \eqn{\frac{1}{n^2}\sum_{i=1}^{n}(X_{i}-\mu)(X_{i}-\mu)'}.
  \item 'gamma', \eqn{\frac{1}{n^2}\sum_{i=1}^{n}(X_{i}-\mu)\xi_{i}}.
  \item 'mean_resid', \eqn{\frac{1}{n}\sum_{i=1}^{n}(X_{i}-\mu)}.
}
}
\description{
AUC Calculation for Augmentation Estimator.
}
