% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AUC.R
\name{AUC}
\alias{AUC}
\title{Find Area Under the Curve}
\usage{
AUC(times, values, tau, subdiv = 10000)
}
\arguments{
\item{times}{Time points.}

\item{values}{Values.}

\item{tau}{Truncation time.}

\item{subdiv}{Subdivisions for integration.}
}
\value{
Numeric area under the curve.
}
\description{
Find Area Under the Curve
}
