% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Augmentation.R
\name{AUC.Stats.Aug}
\alias{AUC.Stats.Aug}
\title{Calculate Test Statistics for Augmentation Estimator.}
\usage{
AUC.Stats.Aug(time, status, arm, idx, covars, tau, alpha, return_areas = FALSE)
}
\arguments{
\item{time}{Observation time.}

\item{status}{Status, coded as 0 for censoring, 1 for event, 2 for death.
Note that subjects who are neither censored nor die are assumed to remain
at risk throughout the observation period.}

\item{arm}{Arm, coded as 1 for treatment, 0 for reference.}

\item{idx}{Unique subject index.}

\item{covars}{Per-subject covariate matrix.}

\item{tau}{Truncation time.}

\item{alpha}{Type I error.}

\item{return_areas}{Return the AUCs?}
}
\value{
If `return_areas`, list containing:
\itemize{
  \item 'marg_area', marginal area for each arm.
  \item 'mcf', mean cumulative function for each strata.
  \item 'contrasts', including the augmented difference of areas.
}
 Else, only 'contrasts' is returned.
}
\description{
Calculate test statistics for augmentation estimator.
}
