% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Stratified.R
\name{AUC.Stats.Strat}
\alias{AUC.Stats.Strat}
\title{Calculate Test Statistics for Stratified Estimator.}
\usage{
AUC.Stats.Strat(
  time,
  status,
  arm,
  idx,
  strata,
  tau,
  alpha,
  return_areas = FALSE
)
}
\arguments{
\item{time}{Observation time.}

\item{status}{Status, coded as 0 for censoring, 1 for event, 2 for death.
Note that subjects who are neither censored nor die are assumed to
remain at risk throughout the observation period.}

\item{arm}{Arm, coded as 1 for treatment, 0 for reference.}

\item{idx}{Unique subject index.}

\item{strata}{Optional stratification factor.}

\item{tau}{Truncation time.}

\item{alpha}{Type I error.}

\item{return_areas}{Return the AUCs?}
}
\value{
If `return_areas`, list containing:
\itemize{
  \item 'avg_mcf', average MCF across strata.
  \item 'contrasts', including the difference and ratio of areas.
  \item 'marg_areas', the marginal areas for each arm.
  \item 'stratum_areas', the per-stratum and per-arm areas.
  \item 'weights', the weights of each stratum.
}
 Else, only 'contrasts' is returned.
}
\description{
Calculate test statistics for stratified estimator.
}
