% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AUC.R
\name{AUCCI}
\alias{AUCCI}
\title{Confidence Intervals for the Difference and Ratio of AUCs.}
\usage{
AUCCI(time, status, arm, idx, tau, B = 2000, alpha = 0.05)
}
\arguments{
\item{time}{Event time.}

\item{status}{Status, coded as 1 for the recurrent event, 0 otherwise.}

\item{arm}{Arm, coded as 1 for treatment, 0 for reference.}

\item{idx}{Subject index.}

\item{tau}{Truncation time.}

\item{B}{Bootstrap replicates.}

\item{alpha}{Alpha level.}
}
\value{
A data.frame containing these columns:
\describe{
  \item{Time}{Truncation time.}
  \item{Arm0}{AUC for arm 0.}
  \item{Arm1}{AUC for arm 1.}
  \item{Contrast}{'A1-A0' for the difference and 'A1/A0' for the ratio.}
  \item{Estimate}{The estimated difference and ratio of AUCs.}
  \item{L}{Confidence lower bound.}
  \item{U}{Confidence upper bound.}
}
}
\description{
Confidence intervals for the difference and ratio of areas under the mean cumulative
count curves, comparing treatment (arm = 1) with reference (arm = 0).
}
