% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Augmentation.R
\name{AugAUC}
\alias{AugAUC}
\title{AUC Calculation for a Single Arm.}
\usage{
AugAUC(data, subj_covars, mu, tau, calc_var = TRUE)
}
\arguments{
\item{data}{Data.frame including: time, status, idx.}

\item{subj_covars}{Per-subject covariate data.}

\item{mu}{Grand mean.}

\item{tau}{Truncation time.}

\item{calc_var}{Calculate analytical variance?}
}
\value{
List containing:
\itemize{
  \item Tabulated mean cumulative function 'mcf'.
  \item Estimated 'area' and its variance 'var_area'.
  \item 'sigma', \eqn{\frac{1}{n^2}\sum_{i=1}^{n}(X_{i}-\mu)(X_{i}-\mu)'}.
  \item 'gamma', \eqn{\frac{1}{n^2}\sum_{i=1}^{n}(X_{i}-\mu)\xi_{i}}.
  \item 'mean_resid', \eqn{\frac{1}{n}\sum_{i=1}^{n}(X_{i}-\mu)}.
}
}
\description{
AUC Calculation for a Single Arm.
}
