% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Mean-cumulative-function.R
\name{AverageMCF}
\alias{AverageMCF}
\title{Calculate Average MCF Curve}
\usage{
AverageMCF(curve_list, weights)
}
\arguments{
\item{curve_list}{List of tabulated MCFs as returned by \code{\link{CalculateMCF}}.}

\item{weights}{Numeric vector of weights.}
}
\value{
Data.frame containing `Time` and the average MCF `Avg_MCF`.
}
\description{
Calculates the weighted average of MCF curves.
}
