% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MCF.R
\name{AvgMCF}
\alias{AvgMCF}
\title{Calculate Average MCF Curve}
\usage{
AvgMCF(curve_list, weights)
}
\arguments{
\item{curve_list}{List of tabulated MCFs as returned by \code{\link{CalcMCF}}.}

\item{weights}{Numeric vector of weights.}
}
\value{
Data.frame containing `Time` and the average MCF `Avg_MCF`.
}
\description{
Calculates the weighted average of MCF curves.
}
