% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{berkley}
\alias{berkley}
\title{UCBAdmissions dataset (Berkeley)}
\format{
A data frame with columns:
\describe{
\item{Dept}{Department (A–F)}
\item{Gender}{Applicant gender (Male, Female)}
\item{Admit}{Admission status (Admitted, Rejected)}
}
}
\source{
\url{https://stat.ethz.ch/R-manual/R-devel/library/datasets/html/UCBAdmissions.html}
}
\usage{
berkley
}
\description{
Admissions information for the six largest departments of the University
of California, Berkeley, in Fall 1973. Frequencies have been expanded so that
each row represents one applicant.
}
\examples{
data(berkley)
head(berkley)

}
\keyword{datasets}
