% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Augmentation.R
\name{Boot.CIs.Aug}
\alias{Boot.CIs.Aug}
\title{Bootstrap Confidence Intervals for Augmentation Estimator.}
\usage{
Boot.CIs.Aug(sim, obs_stats, alpha)
}
\arguments{
\item{sim}{Bootstrap simulation results, as generated by \code{\link{Boot.Sim.Aug}}.}

\item{obs_stats}{Observed contrasts.}

\item{alpha}{Type I error.}
}
\value{
Data.framne containing the equi-talied and highest-density bootstrap
  confidence intervals.
}
\description{
Bootstrap Confidence Intervals for Augmentation Estimator.
}
