% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Main.R
\name{Boot.cis}
\alias{Boot.cis}
\title{Bootstrap Confidence Intervals}
\usage{
Boot.cis(sim, obs_stats, alpha)
}
\arguments{
\item{sim}{Bootstrap simulation results, as generated by \code{\link{Boot.sim}}.}

\item{obs_stats}{Observed contrasts.}

\item{alpha}{Type I error.}
}
\value{
Data.framne containing the equi-talied and highest-density bootstrap
  confidence intervals.
}
\description{
Bootstrap Confidence Intervals
}
