% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Stratified.R
\name{Boot.CIs.Strat}
\alias{Boot.CIs.Strat}
\title{Bootstrap Confidence Intervals}
\usage{
Boot.CIs.Strat(sim, obs_stats, alpha)
}
\arguments{
\item{sim}{Bootstrap simulation results, as generated by \code{\link{Boot.Sim.Strat}}.}

\item{obs_stats}{Observed contrasts.}

\item{alpha}{Type I error.}
}
\value{
Data.frame containing the equi-tailed and highest-density bootstrap
  confidence intervals.
}
\description{
Bootstrap Confidence Intervals
}
