% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Augmentation.R
\name{Boot.Sim.Aug}
\alias{Boot.Sim.Aug}
\title{Bootstrap Inference}
\usage{
Boot.Sim.Aug(time, status, arm, idx, covars, obs_stats, tau, alpha, reps, seed)
}
\arguments{
\item{time}{Observation time.}

\item{status}{Status, coded as 0 for censoring, 1 for event, 2 for death.
Note that subjects who are neither censored nor die are assumed to remain
at risk throughout the observation period.}

\item{arm}{Arm, coded as 1 for treatment, 0 for reference.}

\item{idx}{Unique subject index.}

\item{covars}{Optional covariate matrix. Rows should correspond with the
subject index `idx`. Factor and interaction terms should be expanded.}

\item{obs_stats}{Observed contrasts.}

\item{tau}{Truncation time.}

\item{alpha}{Type I error.}

\item{reps}{Simulations replicates.}

\item{seed}{Simulation seed.}
}
\value{
Data.frame containing:
\itemize{
  \item Bootstrap difference 'boot_diff' and ratio 'boot_ratio' of areas.
  \item An indicator that the bootstrap difference was of the opposite
    sign, '1side_boot_diff'.
}
}
\description{
Constructs bootstrap confidence intervals.
}
