% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Main.R
\name{Boot.sim}
\alias{Boot.sim}
\title{Bootstrap Inference}
\usage{
Boot.sim(data0, data1, obs_stats, tau, alpha, reps, seed)
}
\arguments{
\item{data0}{Data.frame containing `time`, `status`, `idx`, `strata` for arm 0.}

\item{data1}{Data.frame containing `time`, `status`, `idx`, `strata` for arm 1.}

\item{obs_stats}{Observed contrasts.}

\item{tau}{Trunction time.}

\item{alpha}{Type I error.}

\item{reps}{Simulations replicates.}

\item{seed}{Simulation seed.}
}
\value{
Data.frame containing:
\itemize{
  \item Bootstrap difference 'boot_diff' and ratio 'boot_ratio' of areas.
  \item An indicator that the bootstrap difference was of the opposite
    sign, '1side_boot_diff'.
}
}
\description{
Constructs bootstrap confidence intervals.
}
