% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/BootCIs.R
\name{BootCIs}
\alias{BootCIs}
\title{Bootstrap Confidence Intervals for Stratified Estimator.}
\usage{
BootCIs(sim, obs_stats, alpha)
}
\arguments{
\item{sim}{Bootstrap simulation results, as generated by \code{\link{BootSimStrat}}.}

\item{obs_stats}{Observed contrasts.}

\item{alpha}{Type I error.}
}
\value{
Data.frame containing the equi-tailed and highest-density bootstrap
  confidence intervals.
}
\description{
Bootstrap Confidence Intervals for Stratified Estimator.
}
