% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Bootstrap.R
\name{BootData}
\alias{BootData}
\title{Generate Bootstrap Data.frame.}
\usage{
BootData(data, idx_offset = 0)
}
\arguments{
\item{data}{Data.frame.}

\item{idx_offset}{Index offset.}
}
\value{
Bootstrapped data.frame.
}
\description{
Bootstrap observations grouped by an index. The index `idx` is relabeled
in the returned data.frame so that resampled group has a distinct value.
}
