% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Bootstrap.R
\name{BootSimAug}
\alias{BootSimAug}
\title{Bootstrap Inference}
\usage{
BootSimAug(data, obs_stats, alpha, reps, tau)
}
\arguments{
\item{data}{Data.frame containing: idx, time, status, arm, strata.}

\item{obs_stats}{Observed contrasts.}

\item{alpha}{Type I error.}

\item{reps}{Simulations replicates.}

\item{tau}{Truncation time.}
}
\value{
Data.frame containing:
\itemize{
  \item Bootstrap difference 'boot_diff' and ratio 'boot_ratio' of areas.
  \item An indicator that the bootstrap difference was of the opposite
    sign, '1side_boot_diff'.
}
}
\description{
Constructs bootstrap confidence intervals.
}
