\name{btof}
\title{Transform binary data format}
\usage{
btof(
  data)
}
\description{
A function to transform binary input data (i.e., performance-judgement data based on binary input variables) to adjust the letters for that the data can be used as input for the functions `participant_summary`, `calibrationaccuracy`, `miscalibration`, and all plotting functions. 
}
\arguments{
\item{data}{data frame with letters from a to d corresponding to the judgement values}
}
\examples{}
\references{}