% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Augmentation.R
\name{CalcAugAUC}
\alias{CalcAugAUC}
\title{Calculate Test Statistics for Augmentation Estimator.}
\usage{
CalcAugAUC(data, tau, alpha = 0.05, return_areas = FALSE)
}
\arguments{
\item{data}{Data.frame containing (arm, idx, status, time, jump_weights).}

\item{tau}{Truncation time.}

\item{alpha}{Type I error.}

\item{return_areas}{Return the AUCs?}
}
\value{
If return_areas is TRUE, list containing:
\itemize{
  \item 'marg_area', marginal area for each arm.
  \item 'mcf', mean cumulative function for each strata.
  \item 'contrasts', including the augmented difference of areas.
}
}
\description{
Calculate test statistics for augmentation estimator.
}
