% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Permutation.R
\name{CalcP}
\alias{CalcP}
\title{Calculate P}
\usage{
CalcP(p)
}
\arguments{
\item{p}{Vector of 1/0 rejection indicators.}
}
\value{
Numeric p-value.
}
\description{
Calculates p-value from 1-sided rejection indicator.
}
