% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Stratified.R
\name{CalcStratAUC}
\alias{CalcStratAUC}
\title{Calculate Test Statistics for Stratified Estimator.}
\usage{
CalcStratAUC(data, tau, alpha, return_areas = FALSE)
}
\arguments{
\item{data}{Data.frame containing: idx, time, status, arm, strata.}

\item{tau}{Truncation time.}

\item{alpha}{Type I error.}

\item{return_areas}{Return the AUCs?}
}
\value{
If `return_areas`, list containing:
\itemize{
  \item 'avg_mcf', average MCF across strata.
  \item 'contrasts', including the difference and ratio of areas.
  \item 'marg_areas', the marginal areas for each arm.
  \item 'stratum_areas', the per-stratum and per-arm areas.
  \item 'weights', the weights of each stratum.
}
 Else, only 'contrasts' is returned.
}
\description{
Calculate Test Statistics for Stratified Estimator.
}
