% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Mean-cumulative-function.R
\name{CalculateMCF}
\alias{CalculateMCF}
\title{Calculate Mean Cumulative Function}
\usage{
CalculateMCF(time, status, idx)
}
\arguments{
\item{time}{Observation time.}

\item{status}{Status, coded as 0 for censoring, 1 for event, 2 for death.}

\item{idx}{Unique subject index.}
}
\value{
Data.frame with these columns:
\itemize{
  \item `event_times` distinct recurrent event times.
  \item `nar` number at risk at the event time. 
  \item `events` number of events that occurred. 
  \item `surv`, survival probability.
  \item `event_rate`, instantaneous event rate.
  \item `mcf`, mean cumulative function. 
}
}
\description{
Tabulates the mean cumulative function. See equation 2.1 of 
<doi:10.1111/j.0006-341X.2000.00554.x>.
}
