% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AUC.R
\name{CalcVarAUC.i}
\alias{CalcVarAUC.i}
\title{Calculate Variance Contribution of an Individual Subject to Area Under
  the Mean Cumulative Function.}
\usage{
CalcVarAUC.i(mcf, time, status, tau)
}
\arguments{
\item{mcf}{Tabulated MCF as returned by \code{\link{CalcMCF}}.}

\item{time}{Observation times for a single subject.}

\item{status}{Status indicator for a single subject.}

\item{tau}{Truncation time.}
}
\value{
Numeric square influence function for a single observation.
}
\description{
Calculate Variance Contribution of an Individual Subject to Area Under
  the Mean Cumulative Function.
}
