% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AUC.R
\name{CalcVarAUC}
\alias{CalcVarAUC}
\title{Calculate Variance of the Area Under the Mean Cumulative Function}
\usage{
CalcVarAUC(mcf, time, status, idx, tau)
}
\arguments{
\item{mcf}{Tabulated MCF as returned by \code{\link{CalcMCF}}.}

\item{time}{Observation times.}

\item{status}{Status indicators.}

\item{idx}{Unique subject index.}

\item{tau}{Truncation time.}
}
\value{
Numeric estimated variance.
}
\description{
Calculate Variance of the Area Under the Mean Cumulative Function
}
