% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MCF.R
\name{CalcVarMCF.i}
\alias{CalcVarMCF.i}
\title{Calculate MCF Variance Contribution for a Single Subject}
\usage{
CalcVarMCF.i(mcf, time, status)
}
\arguments{
\item{mcf}{Tabulated MCF as returned by \code{\link{CalcMCF}}.}

\item{time}{Observation times for a single subject.}

\item{status}{Status indicator for a single subject.}
}
\value{
Numeric vector containing the squared influence function for a single
  subject at each distinct event event.
}
\description{
Calculate MCF Variance Contribution for a Single Subject
}
