% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/carryForward.R
\name{carryforward}
\alias{carryforward}
\title{Carry forward last non-NA}
\usage{
carryforward(x, na.rm = TRUE)
}
\arguments{
\item{x}{name of variable for which last non NA value is to be repeated}

\item{na.rm}{If x starts with one or more NAs the leading NAs will be removed by default. Set na.rm = FALSE to retain leading NAs}
}
\value{
a variable similar to the input but with non-leading NAs replaced by last non-NA value. Leading non-NAs will either be removed or remain as NAs.
}
\description{
Function to let last non NA value of a variable be carried forward
}
\examples{
x <- c(NA, NA, NA, NA, 1, 3, 52, 5, 6, NA, 532, NA, NA, NA, 43)
y <- carryforward(x)
z <- carryforward(x, na.rm = FALSE)
}
