% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Main.R
\name{CompareAUCs}
\alias{CompareAUCs}
\title{Inference on the Area Under the Cumulative Count Curve}
\usage{
CompareAUCs(
  data,
  alpha = 0.05,
  arm_name = "arm",
  boot = FALSE,
  cens_after_last = TRUE,
  covars = NULL,
  idx_name = "idx",
  perm = FALSE,
  reps = 2000,
  status_name = "status",
  strata = NULL,
  tau = NULL,
  time_name = "time",
  jump_weights = NULL
)
}
\arguments{
\item{data}{Data.frame.}

\item{alpha}{Type I error level.}

\item{arm_name}{Name of column containing treatment arm. Must be coded as 1
for treatment, 0 for reference.}

\item{boot}{Logical, construct bootstrap confidence intervals (slow)?}

\item{cens_after_last}{Should subjects who lack an explicit censoring time
be censored after their last observed event?}

\item{covars}{Optional covariate matrix. Rows should correspond with the
subject index `idx`. Factor and interaction terms should be expanded.}

\item{idx_name}{Name of column containing a unique subject index.}

\item{perm}{Logical, perform permutation test (slow)?}

\item{reps}{Number of replicates for bootstrap/permutation inference.}

\item{status_name}{Name of column containing the status. Must be coded as 0 for
censoring, 1 for event, 2 for death. Each subject should have an 
observation-terminating event, either censoring or death.}

\item{strata}{Optional stratification factor. Should not be provided if a
covariate matrix is provided.}

\item{tau}{Numeric truncation time.}

\item{time_name}{Name of column containing the observation time.}

\item{jump_weights}{Optional column of jump weights, controlling the size of the jump
in the cumulative count curve at times with status == 1.}
}
\value{
Object of class \code{CompStratAUCs} or \code{CompAugAUCs} with these slots:
\itemize{
  \item `@Areas`: The AUC for each arm.
  \item `@CIs`: Observed difference and ratio in areas with confidence intervals.
  \item `@Curves`: Mean cumulative count curve for each arm; averaged across strata
    if present.
  \item `@Pvals`: Bootstrap and permutation p-values.
  \item `@Reps`: Bootstrap and permutation realizations of the test statistics.
  \item `@Weights`: Per-stratum weights and AUCs.
}
}
\description{
Confidence intervals and p-values for the difference and ratio of areas under
the mean cumulative count curves, comparing treatment (arm = 1) with
reference (arm = 0).
}
\details{
Two methods of p-value calculation are available. For 'perm', treatment
assignments are permuted on each iteration, and the p-value is the
proportion of the *null* statistics that are as or more extreme than
the *observed* statistics. For 'boot', the p-value is twice the proportion
of bootstrap replicates on which the sign of the difference is areas is
reversed.
}
\examples{
\donttest{
# Simulate data set.
covariates <- data.frame(arm = c(rep(1, 50), rep(0, 50)))
data <- GenData(
  beta_event = log(0.5),
  covariates = covariates
)

aucs <- CompareAUCs(
  data,
  tau = 2,
  boot = TRUE,
  perm = TRUE,
  reps = 25,
  alpha = 0.05
)
show(aucs)
}
}
