% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Stratified.R
\name{CompareStratAUCs}
\alias{CompareStratAUCs}
\title{Inference on the Area Under the Cumulative Count Curve}
\usage{
CompareStratAUCs(
  data,
  tau,
  alpha = 0.05,
  boot = FALSE,
  perm = FALSE,
  reps = 2000
)
}
\arguments{
\item{data}{Formatted data.frame. \code{\link{FormatData}}.}

\item{tau}{Truncation time.}

\item{alpha}{Alpha level.}

\item{boot}{Logical, construct bootstrap confidence intervals?}

\item{perm}{Logical, perform permutation test?}

\item{reps}{Replicates for bootstrap/permutation inference.}
}
\value{
Object of class \code{CompStratAUCs} with these slots:
\itemize{
  \item `@Areas`: The AUC for each arm.
  \item `@CIs`: Observed difference and ratio in areas with confidence intervals.
  \item `@Curves`: Mean cumulative count curve for each arm; averaged across strata
    if present.
  \item `@Pvals`: Bootstrap and permutation p-values.
  \item `@Reps`: Bootstrap and permutation realizations of the test statistics.
  \item `@Weights`: Per-stratum weights and AUCs.
}
}
\description{
Confidence intervals and p-values for the difference and ratio of areas under
the mean cumulative count curves, comparing treatment (arm = 1) with
reference (arm = 0).
}
\examples{
\donttest{
# Simulate data set.
n <- 100
covariates <- data.frame(
  arm = c(rep(1, n/2), rep(0, n/2)),
  strata = stats::rbinom(n, 1, 0.2)
)
data <- GenData(
  beta_event = c(log(0.5), log(0.8)),
  covariates = covariates
)

aucs <- CompareAUCs(
  data,
  strata = data$strata,
  tau = 2,
  boot = TRUE,
  perm = TRUE,
  reps = 25,
  alpha = 0.05
)
show(aucs)
}
}
