% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Class.R
\docType{class}
\name{compAUCs-class}
\alias{compAUCs-class}
\title{Compare AUCs Object}
\description{
Defines the object class returned by \code{\link{CompareAUCs}}.
}
\section{Slots}{

\describe{
\item{\code{Areas}}{Overall sample sizes and AUCs for each arm.}

\item{\code{CIs}}{Equi-tailed and highest density confidence intervals.}

\item{\code{Curves}}{Mean cumulative function for each treatment arm, averaged across
strata.}

\item{\code{Pvals}}{Bootstrap and permutation p-values.}

\item{\code{Reps}}{Bootstrap and permutation statistics.}

\item{\code{Weights}}{Per-stratum weights and areas.}
}}

