% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Main.R
\name{CompAUCs.perm}
\alias{CompAUCs.perm}
\title{Permutation Inference}
\usage{
CompAUCs.perm(data, obs_stats, tau, alpha, reps, seed)
}
\arguments{
\item{data}{Data.frame containing `time`, `status`, `arm`, `idx`, `strata`.}

\item{obs_stats}{Observed contrasts.}

\item{tau}{Trunction time.}

\item{alpha}{Type I error.}

\item{reps}{Simulations replicates.}

\item{seed}{Simulation seed.}
}
\value{
Data.frame containing:
\itemize{
  \item Permutation difference 'perm_diff' and ratio 'perm_ratio' of areas.
  \item Indicators that the permutation difference or ratio was as or more extreme
    than the observed difference or ratio.
}
}
\description{
Permutation Inference
}
