% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Stratified.R
\name{CompAUCs.Perm.Strat}
\alias{CompAUCs.Perm.Strat}
\title{Permutation Inference}
\usage{
CompAUCs.Perm.Strat(
  time,
  status,
  arm,
  idx,
  strata,
  obs_stats,
  tau,
  alpha,
  reps
)
}
\arguments{
\item{time}{Observation time.}

\item{status}{Status, coded as 0 for censoring, 1 for event, 2 for death.
Note that subjects who are neither censored nor die are assumed to
remain at risk throughout the observation period.}

\item{arm}{Arm, coded as 1 for treatment, 0 for reference.}

\item{idx}{Unique subject index.}

\item{strata}{Optional stratification factor.}

\item{obs_stats}{Observed contrasts.}

\item{tau}{Truncation time.}

\item{alpha}{Type I error.}

\item{reps}{Simulations replicates.}
}
\value{
Data.frame containing:
\itemize{
  \item Permutation difference 'perm_diff' and ratio 'perm_ratio' of areas.
  \item Indicators that the permutation difference or ratio was as or more extreme
    than the observed difference or ratio.
}
}
\description{
Permutation Inference
}
