% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Class.R
\docType{class}
\name{CompAugAUCs-class}
\alias{CompAugAUCs-class}
\title{Compare Augmented AUCs Object}
\description{
Defines the object class returned by \code{\link{CompareAugAUCs}}.
}
\section{Slots}{

\describe{
\item{\code{Areas}}{Area under the MCF for each arm.}

\item{\code{CIs}}{Confidence intervals.}

\item{\code{MCF}}{Tabulated Mean cumulative function for each treatment arm.}

\item{\code{Pvals}}{P-values.}

\item{\code{Reps}}{List of data.frame containing the bootstrap/permutation replicates.}
}}

