% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Class.R
\docType{class}
\name{CompStratAUCs-class}
\alias{CompStratAUCs-class}
\title{Compare Stratified AUCs Object}
\description{
Defines the object class returned by \code{\link{CompareStratAUCs}}.
}
\section{Slots}{

\describe{
\item{\code{StratumAreas}}{Stratum-specific areas for each arm.}

\item{\code{MargAreas}}{Areas for each arm, marginalized across strata.}

\item{\code{CIs}}{Confidence intervals.}

\item{\code{MCF}}{Mean cumulative function for each treatment arm, averaged across
strata.}

\item{\code{Pvals}}{P-values.}

\item{\code{Reps}}{List of data.frame containing the bootstrap/permutation replicates.}
}}

