\name{confidence_plot}
\title{Visualize the confidence accuracy}
\usage{
confidence_plot(
  data, 
  color = blue, 
  plot = "jitter", 
  colorbg = blue202, 
  name = "", 
  export = FALSE)
}
\description{
A function to visualize the confidence accuracy. 
}
\arguments{
\item{data}{vector with the confidence accuracy values per participant}
\item{color}{color of the boxplot}
\item{colorbg}{background color of the violin plot}
\item{plot}{type of plot: either "jitter" or "violin"}
\item{name}{name of the plot to save it}
\item{export}{TRUE/FALSE value to determine whether the plot should be exported}
}
\examples{}
\references{}