% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AUC.R
\name{ContrastAreas}
\alias{ContrastAreas}
\title{Contrast Summary Statistics}
\usage{
ContrastAreas(area1, area0, se1, se0, alpha)
}
\arguments{
\item{area1}{Statistic for arm 1.}

\item{area0}{Statistic for arm 0.}

\item{se1}{Standard error for arm 1.}

\item{se0}{Standard error for arm 0.}

\item{alpha}{Type I error.}
}
\value{
Data.frame containing:
\itemize{
  \item 'Contrast' and estimate 'Est'.
  \item Lower 'L' and upper 'U' confidence bounds.
  \item 'P' value.
}
}
\description{
Finds the difference and ratio of summary statistics, comparing two arms.
}
