% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utilities.R
\name{convert_to_frequency_format}
\alias{convert_to_frequency_format}
\title{Convert Individual-Level Data to Frequency Format}
\usage{
convert_to_frequency_format(data)
}
\arguments{
\item{data}{Data frame with categorical variables}
}
\value{
A data frame containing all unique combinations of variable values,
along with a frequency column \code{y} giving the count for each combination.
}
\description{
Aggregates individual observations into frequency counts for each unique
combination of variable levels
}
\examples{
df <- data.frame(A = c(1,1,2,2), B = c("x","x","y","y"))
convert_to_frequency_format(df)

}
