% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AUC.R
\name{FindAUC}
\alias{FindAUC}
\title{Find Area Under the Mean Cumulative Count Curve.}
\usage{
FindAUC(times, values, tau)
}
\arguments{
\item{times}{MCF times.}

\item{values}{MCF values}

\item{tau}{Truncation time.}
}
\value{
Numeric area under the curve.
}
\description{
Find Area Under the Mean Cumulative Count Curve.
}
