% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Format.R
\name{FormatSubj}
\alias{FormatSubj}
\title{Format Data for a Single Subject}
\usage{
FormatSubj(df)
}
\arguments{
\item{df}{Data.frame for a single subject.}
}
\value{
Data.frame with an added censoring event.
}
\description{
Format Data for a Single Subject
}
