% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Resampling.R
\name{GroupBoot}
\alias{GroupBoot}
\title{Grouped Bootstrap.}
\usage{
GroupBoot(data)
}
\arguments{
\item{data}{Data.frame containing `idx`.}
}
\value{
Bootstrapped data.frame.
}
\description{
Bootstrap observations grouped by an index.
}
