% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CIs.R
\name{HighDensCI}
\alias{HighDensCI}
\title{Highest Density Confidence Interval}
\usage{
HighDensCI(x, alpha = 0.05, min_tail_prob = 0.005, intervals = 100)
}
\arguments{
\item{x}{Bootstrap replicates.}

\item{alpha}{Alpha level.}

\item{min_tail_prob}{Minimum alpha per tail.}

\item{intervals}{Intervals to consider.}
}
\description{
Highest Density Confidence Interval
}
