% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AUC.R
\name{MargAUC}
\alias{MargAUC}
\title{Calculate Marginal Area}
\usage{
MargAUC(areas, ses, weights)
}
\arguments{
\item{areas}{Estimated statistics.}

\item{ses}{Standard errors.}

\item{weights}{Weights.}
}
\value{
Data.frame containing:
\itemize{
  \item The marginal area.
  \item The standard error of the area.
}
}
\description{
Calculate Marginal Area
}
