% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MCF.R
\name{Martingales}
\alias{Martingales}
\title{Calculate Martingales for a Single Subject}
\usage{
Martingales(mcf, time, status)
}
\arguments{
\item{mcf}{Tabulated MCF as returned by \code{\link{CalcMCF}}.}

\item{time}{Observation times for a single subject.}

\item{status}{Status indicator for a single subject.}
}
\value{
Data.frame containing:
\itemize{
  \item Distinct observation times from `mcf`.
  \item 'censor', 'death', and 'event' indicators.
  \item 'nar', an at-risk indicator.
  \item 'dM_event' and 'dM_death', increments of the event and 
    death martingales.
}
}
\description{
Calculate Martingales for a Single Subject
}
