% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/PlotMCF1.R
\name{MCFCurve}
\alias{MCFCurve}
\title{MCF Curve}
\usage{
MCFCurve(
  data,
  idx_name = "idx",
  status_name = "status",
  time_name = "time",
  jump_weights = NULL
)
}
\arguments{
\item{data}{Data.frame.}

\item{idx_name}{Name of index (subject identifier) column in data.}

\item{status_name}{Name of status column in data.}

\item{time_name}{Name of column column in data.}

\item{jump_weights}{Optional column of jump weights, controlling the size of the jump
in the cumulative count curve at times with status == 1.}
}
\value{
stepfun.
}
\description{
Construct a function to evaluate the mean cumulative function at a given time
for a single treatment arm.
}
