% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/minmax.R
\name{minmax}
\alias{minmax}
\title{Longest shared follow-up}
\usage{
minmax(data)
}
\arguments{
\item{data}{data.frame from the MCC-function with grp-parameter specified (i.e. with variable grp)}
}
\value{
numeric value of the longest time observed in all groups
}
\description{
Function to identify the longest follow-up in the group with shortest follow-up
}
\examples{
mccOutput <- mcc(data = mcc_events,
                 idvar = "id",
                 timevar = "time",
                 eventvar = "event",
                 grpvar = "grp",
                 weightvar = "weight")
minmax(mccOutput)
}
