\name{miscalibration_plot}
\title{Visualize the miscalibration effect}
\usage{
miscalibration_plot(
  miscalibration, 
  calibrationaccuracy, 
  export = TRUE, 
  name = "")
}
\description{
A function to visualize the miscalibration effect.
}
\arguments{
\item{miscalibration}{vector with the miscalibration values per participant}
\item{calibrationaccuracy}{vector with the calibration accuracy values per participant}
\item{name}{name of the plot to save it}
\item{export}{TRUE/FALSE value to determine whether the plot should be exported}
}
\examples{}
\references{}