% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/PlotMvMCF.R
\name{MvMCFCurve}
\alias{MvMCFCurve}
\title{Multivariate MCF Curves}
\usage{
MvMCFCurve(
  data,
  cens_after_last = TRUE,
  event_type_name = "event_type",
  idx_name = "idx",
  status_name = "status",
  time_name = "time"
)
}
\arguments{
\item{data}{Data.frame in long format with an \code{event_type} column.}

\item{cens_after_last}{Censor subjects without a terminal row after their
last recurrent event?}

\item{event_type_name}{Name of event type column.}

\item{idx_name}{Name of subject index column.}

\item{status_name}{Name of status column.}

\item{time_name}{Name of time column.}
}
\value{
Named list of \code{stepfun} objects, one per event type.
}
\description{
Return stepfun objects evaluating the mean cumulative function at each
event type for one sample.
}
