% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/PlotMCF.R
\name{NARCurve}
\alias{NARCurve}
\title{Number at Risk Curve}
\usage{
NARCurve(data, idx_name = "idx", status_name = "status", time_name = "time")
}
\arguments{
\item{data}{Data.frame.}

\item{idx_name}{Name of index (subject identifier) column in data.}

\item{status_name}{Name of status column in data.}

\item{time_name}{Name of column column in data.}
}
\value{
stepfun.
}
\description{
Return a function that calculates the number at risk for a single
treatment arm.
}
