% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utilities.R
\name{perform_mode_shift}
\alias{perform_mode_shift}
\title{Perform Mode-Shift Algorithm}
\usage{
perform_mode_shift(edge_list, n_nodes)
}
\arguments{
\item{edge_list}{A data frame with V1, V2 columns defining node connections}

\item{n_nodes}{Integer specifying the total number of nodes in the graph}
}
\value{
Vector of cluster assignments.
}
\description{
Iteratively reassigns each node to its most connected higher-density neighbor
until convergence. The final assignments identify modal regions in the density.
}
\keyword{internal}
