% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Bootstrap.R
\name{PermData}
\alias{PermData}
\title{Generate Permuted Data.frame.}
\usage{
PermData(data)
}
\arguments{
\item{data}{Data.frame.}
}
\value{
Bootstrapped data.frame.
}
\description{
Permute observations grouped by an index by flipping the treatment arm
`arm` of an observations with probability 0.5. Arm is assumed to be labeled
1 for treatment, 0 for reference.
}
