% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Permutation.R
\name{PermSimStrat}
\alias{PermSimStrat}
\title{Permutation Inference for Stratified Estimator.}
\usage{
PermSimStrat(data, obs_stats, tau, alpha, reps)
}
\arguments{
\item{data}{Data.frame containing: idx, time, status, arm, strata.}

\item{obs_stats}{Observed contrasts.}

\item{tau}{Truncation time.}

\item{alpha}{Type I error.}

\item{reps}{Simulations replicates.}
}
\value{
Data.frame containing:
\itemize{
  \item Permutation difference 'perm_diff' and ratio 'perm_ratio' of areas.
  \item Indicators that the permutation difference or ratio was as or more extreme
    than the observed difference or ratio.
}
}
\description{
Permutation Inference for Stratified Estimator.
}
