% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/PlotMCF2.R
\name{PlotNARs}
\alias{PlotNARs}
\title{Plot Two Sample Number at Risk}
\usage{
PlotNARs(
  data,
  x_breaks,
  arm_name = "arm",
  idx_name = "idx",
  status_name = "status",
  time_name = "time",
  x_labs = NULL,
  x_max = NULL,
  x_name = NULL,
  y_labs = c("Ctrl", "Trt")
)
}
\arguments{
\item{data}{Data.frame.}

\item{x_breaks}{X-axis breaks.}

\item{arm_name}{Name of arm column.}

\item{idx_name}{Name of index (subject identifier) column in data.}

\item{status_name}{Name of status column.}

\item{time_name}{Name of time column.}

\item{x_labs}{X-axis tick labels.}

\item{x_max}{X-axis upper limit.}

\item{x_name}{X-axis label.}

\item{y_labs}{Y-axis tick labels.}
}
\value{
ggplot.
}
\description{
Deprecated: use \code{\link{PlotMCFs}} with \code{show_mcf = FALSE} and
\code{show_nar = TRUE}.
}
