% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/PlotMCF1.R
\name{PlotOneSampleMCF}
\alias{PlotOneSampleMCF}
\title{Plot One Sample Mean Cumulative Function}
\usage{
PlotOneSampleMCF(
  data,
  color = "#C65842",
  color_lab = "Arm",
  idx_name = "idx",
  status_name = "status",
  tau = NULL,
  time_name = "time",
  title = NULL,
  jump_weights = NULL,
  x_breaks = NULL,
  x_lim = NULL,
  x_name = "Time",
  y_breaks = NULL,
  y_lim = NULL,
  y_name = "Mean Cumulative Count"
)
}
\arguments{
\item{data}{Data.frame.}

\item{color}{Line color when a single arm is plotted.}

\item{color_lab}{Legend label when a single arm is plotted.}

\item{idx_name}{Name of index (subject identifier) column in data.}

\item{status_name}{Name of status column in data.}

\item{tau}{Truncation time.}

\item{time_name}{Name of time column in data.}

\item{title}{Plot title.}

\item{jump_weights}{Optional column of jump weights for one-arm MCF jumps.}

\item{x_breaks}{X-axis breaks for the NAR row.}

\item{x_lim}{X-axis limits.}

\item{x_name}{X-axis label.}

\item{y_breaks}{Y-axis breaks.}

\item{y_lim}{Y-axis limits.}

\item{y_name}{Y-axis label.}
}
\value{
ggplot object.
}
\description{
Deprecated: use \code{\link{PlotMCFs}} with \code{show_nar = FALSE} and
\code{show_auc = FALSE}.
}
