% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/PlotMCF1.R
\name{PlotOneSampleNAR}
\alias{PlotOneSampleNAR}
\title{Plot One Sample Number at Risk}
\usage{
PlotOneSampleNAR(
  data,
  x_breaks,
  idx_name = "idx",
  status_name = "status",
  time_name = "time",
  x_labs = NULL,
  x_max = NULL,
  x_name = NULL,
  y_lab = "Arm"
)
}
\arguments{
\item{data}{Data.frame.}

\item{x_breaks}{X-axis breaks.}

\item{idx_name}{Name of index (subject identifier) column in data.}

\item{status_name}{Name of status column.}

\item{time_name}{Name of time column.}

\item{x_labs}{X-axis tick labels.}

\item{x_max}{X-axis upper limit.}

\item{x_name}{X-axis label.}

\item{y_lab}{Y-axis tick label.}
}
\value{
ggplot.
}
\description{
Deprecated: use \code{\link{PlotMCFs}} with \code{show_mcf = FALSE} and
\code{show_nar = TRUE}.
}
