% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Class.R
\name{print.compAUCs}
\alias{print.compAUCs}
\title{Print Method for Compre AUCs Object.}
\usage{
\method{print}{compAUCs}(x, ...)
}
\arguments{
\item{x}{An object of class \code{compAUCs}.}

\item{...}{Unused.}
}
\description{
Print method for objects of class \code{compAUCs}.
}
