% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Class.R
\name{print.CompAugAUCs}
\alias{print.CompAugAUCs}
\title{Print Method for Compare Augmented AUCs Object.}
\usage{
\method{print}{CompAugAUCs}(x, ...)
}
\arguments{
\item{x}{An object of class \code{CompAugAUCs}.}

\item{...}{Unused.}
}
\description{
Print method for objects of class \code{CompAugAUCs}.
}
