% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Class.R
\name{print.CompMvAugAUCs}
\alias{print.CompMvAugAUCs}
\title{Print Method for Compare Multivariate Augmented AUCs Object.}
\usage{
\method{print}{CompMvAugAUCs}(x, ...)
}
\arguments{
\item{x}{An object of class \code{CompMvAugAUCs}.}

\item{...}{Unused.}
}
\description{
Print method for objects of class \code{CompMvAugAUCs}.
}
