% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Class.R
\name{print.CompStratAUCs}
\alias{print.CompStratAUCs}
\title{Print Method for Compare Stratified AUCs Object.}
\usage{
\method{print}{CompStratAUCs}(x, ...)
}
\arguments{
\item{x}{An object of class \code{CompareStratAUCs}.}

\item{...}{Unused.}
}
\description{
Print method for objects of class \code{CompareStratAUCs}.
}
