% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AUC.R
\name{PsiAUCi}
\alias{PsiAUCi}
\title{Calculate Influence Function Contribution of a Single Subject.}
\usage{
PsiAUCi(mcf, time, status, tau)
}
\arguments{
\item{mcf}{Tabulated MCF as returned by \code{\link{CalcMCF}}.}

\item{time}{Observation times for a single subject.}

\item{status}{Status indicator for a single subject.}

\item{tau}{Truncation time.}
}
\value{
Numeric influence function for a single observation.
}
\description{
Calculate Influence Function Contribution of a Single Subject.
}
