% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{SimDataCpp}
\alias{SimDataCpp}
\title{Simulate Data for Multiple Subjects}
\usage{
SimDataCpp(censoring_rate, death_rate, idx, event_rate, tau)
}
\arguments{
\item{censoring_rate}{Rate of censoring.}

\item{death_rate}{Rate of terminal events.}

\item{idx}{Subject index.}

\item{event_rate}{Rate of events.}

\item{tau}{Truncation time.}
}
\value{
Recurrent event data for a single subject.
}
\description{
Simulate Data for Multiple Subjects
}
