% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{SimMvDataCpp}
\alias{SimMvDataCpp}
\title{Simulate multivariate recurrent event data for multiple subjects.}
\usage{
SimMvDataCpp(censoring_rate, death_rate, idx, event_rates, tau)
}
\arguments{
\item{censoring_rate}{Rate of censoring.}

\item{death_rate}{Rate of terminal events.}

\item{idx}{Subject index.}

\item{event_rates}{Matrix of recurrent event rates (n subjects x K types).}

\item{tau}{Truncation time.}
}
\value{
Long-format recurrent event data with event_type.
}
\description{
Simulate multivariate recurrent event data for multiple subjects.
}
