% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Stratified.R
\name{SingleArmStratAUC}
\alias{SingleArmStratAUC}
\title{Single-Arm Area Under the Cumulative Count Curve}
\usage{
SingleArmStratAUC(data, tau, alpha = 0.05, boot = FALSE, reps = 2000)
}
\arguments{
\item{data}{Formatted data.frame. \code{\link{FormatData}}.}

\item{tau}{Truncation time.}

\item{alpha}{Alpha level.}

\item{boot}{Logical, construct bootstrap confidence intervals?}

\item{reps}{Replicates for bootstrap/permutation inference.}
}
\value{
Object of class compAUCs with these slots:
\itemize{
  \item `@Areas`: The AUC for each arm.
  \item `@CIs`: Observed difference and ratio in areas with confidence intervals.
  \item `@Curves`: Mean cumulative count curve for each arm; averaged across strata
    if present.
  \item `@Pvals`: Bootstrap and permutation p-values.
  \item `@Reps`: Bootstrap and permutation realizations of the test statistics.
  \item `@Weights`: Per-stratum weights and AUCs.
}
}
\description{
Confidence intervals and p-values for the difference and ratio of areas under
the mean cumulative count curves, comparing treatment (arm = 1) with
reference (arm = 0).
}
