% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{titanic}
\alias{titanic}
\title{Titanic Dataset}
\format{
A data frame with columns:
\describe{
\item{Class}{Passenger class (1st, 2nd, 3rd, Crew)}
\item{Sex}{Gender (Male, Female)}
\item{Survived}{Survival status (Yes, No)}
}
}
\usage{
titanic
}
\description{
Classical Titanic passenger data containing information on survival
status according to Class and Sex. Frequencies have been expanded so that
each row represents one passenger.
}
\examples{
data(titanic)
head(titanic)

}
\keyword{datasets}
