% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/PlotMCF2.R
\name{TwoSampleNARFrame}
\alias{TwoSampleNARFrame}
\title{Two Sample Number at Risk Plotting Frame}
\usage{
TwoSampleNARFrame(
  data,
  x_breaks,
  arm_name = "arm",
  idx_name = "idx",
  status_name = "status",
  time_name = "time"
)
}
\arguments{
\item{data}{Data.frame.}

\item{x_breaks}{Time points at which to determine the NARs.}

\item{arm_name}{Name of arm column.}

\item{idx_name}{Name of index (subject identifier) column in data.}

\item{status_name}{Name of status column.}

\item{time_name}{Name of time column.}
}
\value{
Data.frame containing `time`, `nar_ctrl`, `nar_trt`.
}
\description{
Two sample numbers at risk for recurrent events data.
}
