% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AUC.R
\name{AUC.Area}
\alias{AUC.Area}
\title{Calculate AUC}
\usage{
AUC.Area(time, status, idx, tau, calc_var = TRUE)
}
\arguments{
\item{time}{Observation time.}

\item{status}{Status, coded as 0 for censoring, 1 for event, 2 for death. 
Note that subjects who are neither censored nor die are assumed to
remain at risk throughout the observation period.}

\item{idx}{Unique subject index.}

\item{tau}{Truncation time.}

\item{calc_var}{Calculate analytical variance?}
}
\value{
Data.frame containing:
\itemize{
  \item Truncation time 'tau' and estimated 'area'.
  \item Variance 'var_area' and standard error 'se_area', if requested.
}
}
\description{
Used to find the AU MCF for each stratum.
}
