% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AUC.R
\name{AUC.Area}
\alias{AUC.Area}
\title{Calculate MCF and Find Area}
\usage{
AUC.Area(time, status, idx, tau)
}
\arguments{
\item{time}{Observation time.}

\item{status}{Status, coded as 0 for censoring, 1 for event, 2 for death. 
Note that subjects who are neither censored nor die are assumed to
remain at risk throughout the observation period.}

\item{idx}{Unique subject index.}

\item{tau}{Truncation time.}
}
\value{
Numeric AUC.
}
\description{
Calculate MCF and Find Area
}
