% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AUC.R
\name{AUC.Stats}
\alias{AUC.Stats}
\title{Calculate Test Statistics}
\usage{
AUC.Stats(data, tau, return_areas = FALSE)
}
\arguments{
\item{data}{Data.frame containing `time`, `status`, `arm`, 
and subject `idx`.}

\item{tau}{Truncation time.}

\item{return_areas}{Return the AUCs?}
}
\value{
If `return_areas = TRUE`, list containing:
\itemize{
  \item `areas` under the mean cumulative count curve for each arm.
  \item `stats`, including the difference and ratio of areas.
}
 If `return_areas = FALSE`, only `stats` is returned.
}
\description{
Calculate Test Statistics
}
