% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AvgMCF.R
\name{CalcMargMCF}
\alias{CalcMargMCF}
\title{Calculate Marginal MCF}
\usage{
CalcMargMCF(data)
}
\arguments{
\item{data}{Data.frame containing {arm, idx, status, strata, time, weights}.}
}
\value{
Data.frame.
}
\description{
Calculates the marginal MCF, averaged across strata, with stratum
weights proportional to the total number of subjects (across arms)
belonging to that stratum.
}
