% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MCF.R
\name{CalcMCF}
\alias{CalcMCF}
\title{Calculate Mean Cumulative Function}
\usage{
CalcMCF(idx, status, time, calc_var = TRUE, weights = NULL)
}
\arguments{
\item{idx}{Unique subject index.}

\item{status}{Status, coded as 0 for censoring, 1 for event, 2 for death.}

\item{time}{Observation time.}

\item{calc_var}{Calculate variance of the MCF?}

\item{weights}{Jump weights.}
}
\value{
Data.frame with these columns:
\itemize{
   \item `times`, distinct observation times.
   \item `censor`, number of censorings.
   \item `death`, number of deaths.
   \item `event`, number of events.
   \item `haz`, instantaneous hazard (of death).
   \item `surv`, survival probability.
   \item `event_rate`, instantaneous event rate.
   \item `mcf`, mean cumulative function. 
   \item `se_mcf`, standard error of the MCF.
}
}
\description{
Tabulates the mean cumulative function. See equation 2.1 of 
 <doi:10.1111/j.0006-341X.2000.00554.x>.
}
