% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AUC.R
\name{CompareAUCs}
\alias{CompareAUCs}
\title{Bootstrap Inference on the Area Under the Cumulative Count Curve}
\usage{
CompareAUCs(
  time,
  status,
  arm,
  idx,
  tau,
  strata = NULL,
  reps = 2000,
  alpha = 0.05,
  pval_type = "2-sided"
)
}
\arguments{
\item{time}{Observation time.}

\item{status}{Status, coded as 0 for censoring, 1 for event, 2 for death. 
Note that subjects who are neither censored nor die are assumed to
remain at risk throughout the observation period.}

\item{arm}{Arm, coded as 1 for treatment, 0 for reference.}

\item{idx}{Unique subject index.}

\item{tau}{Truncation time.}

\item{strata}{Optional stratification factor.}

\item{reps}{Bootstrap replicates.}

\item{alpha}{Alpha level.}

\item{pval_type}{Either '2-sided' or '1-sided'.}
}
\value{
Data.frame containing these columns:
\describe{
  \item{Time}{Truncation time.}
  \item{Arm0}{AUC for arm 0.}
  \item{Arm1}{AUC for arm 1.}
  \item{Contrast}{'A1-A0' for the difference and 'A1/A0' for the ratio.}
  \item{Estimate}{The estimated difference and ratio of AUCs.}
  \item{L}{Confidence lower bound.}
  \item{U}{Confidence upper bound.}
  \item{P}{P-value.}
}
}
\description{
Confidence intervals and p-values for the difference and ratio of areas under
the mean cumulative count curves, comparing treatment (arm = 1) with
reference (arm = 0).
}
\details{
Two methods of p-value calculation are available. For '2-sided', treatment 
assignments are permuted on each bootstrap iteration, and the p-value is the
proportion of the *null* test statistics that are as or more extreme than
the *observed* test statistics. For '1-sided', the p-value is the proportion
of bootstrap replicates on which the sign of the difference is areas is 
reversed.
}
\examples{
\donttest{
data(mcc_data)
set.seed(100)
aucs <- CompareAUCs(
  time = mcc_data$time,
  status = mcc_data$status,
  arm = mcc_data$arm,
  idx = mcc_data$idx,
  tau = 60,
  strata = mcc_data$strata,
  reps = 100,
  alpha = 0.05,
  pval_type = '2-sided'
)
show(aucs)
}
}
