% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Augmentation.R
\name{CompareAugAUCs}
\alias{CompareAugAUCs}
\title{Inference on the Area Under the Cumulative Count Curve}
\usage{
CompareAugAUCs(
  data,
  tau,
  alpha = 0.05,
  boot = FALSE,
  perm = FALSE,
  reps = 2000
)
}
\arguments{
\item{data}{Formatted data frame. \code{\link{FormatData}}.}

\item{tau}{Truncation time.}

\item{alpha}{Alpha level.}

\item{boot}{Logical, construct bootstrap confidence intervals?}

\item{perm}{Logical, perform permutation test?}

\item{reps}{Replicates for bootstrap/permutation inference.}
}
\value{
Object of class \code{CompAugAUCs} with these slots:
\itemize{
  \item `@Areas`: Marginal AUC for each arm.
  \item `@CIs`: Observed difference and ratio in areas with confidence intervals.
  \item `@MCF`: Mean cumulative count curve for each arm.
  \item `@Pvals`: Bootstrap and permutation p-values.
  \item `@Reps`: Bootstrap and permutation realizations of the test statistics.
}
}
\description{
Confidence intervals and p-values for the difference and ratio of areas under
the mean cumulative count curves, comparing treatment (arm = 1) with
reference (arm = 0) with adjusted for covariates.
}
\examples{
\donttest{
# Simulate data set.
n <- 100
covariates <- data.frame(
  arm = c(rep(1, n/2), rep(0, n/2)),
  covar = rnorm(n)
)
data <- GenData(
  beta_event = c(log(0.5), 1),
  covariates = covariates
)

aucs <- CompareAUCs(
  data,
  covars = data$covar,
  tau = 2,
  boot = TRUE,
  perm = TRUE,
  reps = 25,
  alpha = 0.05
)
show(aucs)
}
}
