% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Class.R
\docType{class}
\name{compAUCs-class}
\alias{compAUCs-class}
\title{Compare AUCs Object}
\description{
Defines the object class returned by \code{\link{CompareAUCs}}.
}
\section{Slots}{

\describe{
\item{\code{StratumAreas}}{Stratum-specific areas for each arm.}

\item{\code{MargAreas}}{Areas for each arm, marginalized across strata.}

\item{\code{CIs}}{Confidence intervals.}

\item{\code{MCF}}{Mean cumulative function for each treatment arm, averaged across
strata.}

\item{\code{Pvals}}{P-values.}

\item{\code{Reps}}{List of data.frame containing the bootstrap/permutation replicates.}

\item{\code{Weights}}{Per-stratum weights and areas.}
}}

