% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Contrast.R
\name{ContrastAreas}
\alias{ContrastAreas}
\title{Contrast AUCs}
\usage{
ContrastAreas(marg_areas, alpha = 0.05)
}
\arguments{
\item{marg_areas}{Marginal areas, including area and SE for each arm.}

\item{alpha}{Type I error.}
}
\value{
Data.frame containing the difference and ratio of areas, together
  with the confidence interval and p-value.
}
\description{
Calculate the difference and ratio of areas under the cumulative count curve,
comparing two treatment arms.
}
