% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Class.R
\name{show,compAUCs-method}
\alias{show,compAUCs-method}
\title{Show Method for Compare AUCs Object}
\usage{
\S4method{show}{compAUCs}(object)
}
\arguments{
\item{object}{An object of class \code{compAUCs}.}
}
\description{
Show Method for Compare AUCs Object
}
