% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Class.R
\name{show,CompAugAUCs-method}
\alias{show,CompAugAUCs-method}
\alias{show,CompStratAUCs-method}
\title{Show Method for Compare Augmented AUCs Object}
\usage{
\S4method{show}{CompAugAUCs}(object)

\S4method{show}{CompStratAUCs}(object)
}
\arguments{
\item{object}{An object of class \code{CompareStratAUCs}.}
}
\description{
Show Method for Compare Augmented AUCs Object

Show Method for Compare AUCs Object
}
