% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Format.R
\name{FormatData}
\alias{FormatData}
\title{Format Data}
\usage{
FormatData(idx, time, status, arm, covars, strata)
}
\arguments{
\item{idx}{Unique subject index.}

\item{time}{Observation time.}

\item{status}{Status, coded as 0 for censoring, 1 for event, 2 for death.
Note that subjects who are neither censored nor die are assumed to
remain at risk throughout the observation period.}

\item{arm}{Arm, coded as 1 for treatment, 0 for reference.}

\item{covars}{Optional covariate matrix. Rows should correspond with the
subject index `idx`. Factor and interaction terms should be expanded.}

\item{strata}{Optional stratification factor.}
}
\description{
Format Data
}
