% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Format.R
\name{FormatData}
\alias{FormatData}
\title{Format Data}
\usage{
FormatData(
  data,
  arm_name = "arm",
  cens_after_last = TRUE,
  covars = NULL,
  idx_name = "idx",
  keep_orig_idx = FALSE,
  status_name = "status",
  strata = NULL,
  time_name = "time",
  weights = NULL
)
}
\arguments{
\item{data}{Data.frame.}

\item{arm_name}{Name of column containing treatment arm. Must be coded as 1
for treatment, 0 for reference.}

\item{cens_after_last}{Should subjects who lack an explicit censoring time
be censored after their last observed event?}

\item{covars}{Optional covariate matrix. Rows should correspond with the
subject index `idx`. Factor and interaction terms should be expanded.}

\item{idx_name}{Name of column containing a unique subject index.}

\item{keep_orig_idx}{Keep the original index?}

\item{status_name}{Name of column containing the status. Must be coded as 0 for
censoring, 1 for event, 2 for death. Each subject should have an 
observation-terminating event, either censoring or death.}

\item{strata}{Optional stratification factor. Should not be provided if a
covariate matrix is provided.}

\item{time_name}{Name of column containing the observation time.}

\item{weights}{Optional column of weights.}
}
\value{
Formatted data.frame.
}
\description{
Subsets the key columns for analysis, converts the index to an integer,
adds placeholders for `strata` and `weights` if not specified, checks
to ensure each unique index has 1 and only 1 o
}
