% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Data.R
\name{GenData}
\alias{GenData}
\title{Simulate Recurrent Events Data}
\usage{
GenData(
  base_death_rate = 0.25,
  base_event_rate = 1,
  beta_death = NULL,
  beta_event = NULL,
  censoring_rate = 0.25,
  covariates = NULL,
  frailty_variance = 0,
  min_death_rate = 0.05,
  min_event_rate = 0.05,
  n = 100,
  tau = 4
)
}
\arguments{
\item{base_death_rate}{Baseline arrival rate for the terminal event.}

\item{base_event_rate}{Baseline arrival rate for recurrent events.}

\item{beta_death}{Numeric vector of log rate ratios for the death rate.}

\item{beta_event}{Numeric vector of log rate ratios for the event rate.}

\item{censoring_rate}{Arrival rate for the censoring time.}

\item{covariates}{Numeric design matrix.}

\item{frailty_variance}{Variance of the gamma frailty.}

\item{min_death_rate}{Minimum subject-specific event rate. Must be positive.}

\item{min_event_rate}{Minimum subject-specific event rate. Must be positive.}

\item{n}{Number of subjects. Overwritten by `nrow(covariates)` if covariates are provided.}

\item{tau}{Truncation time.}
}
\value{
Data.frame, containing:
\itemize{
 \item The subject identifier `idx` (index).
 \item `time` and `status` of the event: 0 for censoring, 1 for an 
    event, 2 for death.
 \item The `true_death_rate`, `true_event_rate`, and `frailty`, which are subject-specific.
}
}
\description{
Simulate recurrent events data using exponential censoring, gap, and death 
times. Status is coded as 0 for censoring, 1 for event, 2 for death.
\itemize{
  \item The subject-specific death rate is calculated as frailty x base_death_rate x 
  exp(beta_death x covariates).
  \item The subject-specific event rate is calculated as frailty x base_event_rate x
  exp(beta_event x covariates).
}
}
