% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Data.R
\name{GenData}
\alias{GenData}
\title{Simulate Recurrent Events Data}
\usage{
GenData(n, censoring_rate, event_rate, death_rate, tau)
}
\arguments{
\item{n}{Numbers of subjects.}

\item{censoring_rate}{Rate of censoring.}

\item{event_rate}{Rate of events.}

\item{death_rate}{Rate of terminal events.}

\item{tau}{Truncation time.}
}
\value{
Data.frame containing:
\itemize{
  \item Subject 'idx'.
  \item Observation 'time'.
  \item Observation 'status', 0 for censoring, 1 for event, 2 for death.
}
}
\description{
Simulate recurrent events data using exponential censoring, gap, and death 
times. Status is coded as 0 for censoring, 1 for event, 2 for death.
}
