% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Pseudovalues.R
\name{GenPseudo}
\alias{GenPseudo}
\title{Generate Pseudovalues}
\usage{
GenPseudo(
  data,
  cens_after_last = TRUE,
  idx_name = "idx",
  status_name = "status",
  tau = NULL,
  time_name = "time",
  weights = NULL
)
}
\arguments{
\item{data}{Data.frame.}

\item{cens_after_last}{Should subjects who lack an explicit censoring time
be censored after their last observed event?}

\item{idx_name}{Name of column containing a unique subject index.}

\item{status_name}{Name of column containing the status. Must be coded as 0 for
censoring, 1 for event, 2 for death. Each subject should have an 
observation-terminating event, either censoring or death.}

\item{tau}{Numeric truncation time.}

\item{time_name}{Name of column containing the observation time.}

\item{weights}{Optional column of weights, controlling the size of the jump
in the cumulative count curve at times with status == 1.}
}
\value{
Data.frame with the influence and pseudo-value for each unique observation.
}
\description{
Generate Pseudovalues
}
