% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Resampling.R
\name{GroupBoot}
\alias{GroupBoot}
\title{Grouped Bootstrap.}
\usage{
GroupBoot(data, idx_offset = 0)
}
\arguments{
\item{data}{Data.frame containing `idx`.}

\item{idx_offset}{Index offset.}
}
\value{
Bootstrapped data.frame.
}
\description{
Bootstrap observations grouped by an index. The index `idx` is relabeled
in the returned data.frame so that each resampled group has a distinct value.
}
