% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Stratified.R
\name{MargAUC}
\alias{MargAUC}
\title{Calculate Marginal Area}
\usage{
MargAUC(areas, ses, weights, n)
}
\arguments{
\item{areas}{Estimated statistics.}

\item{ses}{Standard errors.}

\item{weights}{Weights.}

\item{n}{Sample sizes.}
}
\value{
Data.frame containing:
\itemize{
  \item The marginal area.
  \item The standard error of the area.
}
}
\description{
Calculates the marginal AUC across strata.
}
