% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MCC-package.R
\docType{package}
\name{MCC-package}
\alias{MCC}
\alias{MCC-package}
\title{MCC}
\description{
A package to compute mean cumulative count of a recurrent event
}
\seealso{
Useful links:
\itemize{
  \item \url{https://github.com/tjuheide/MCC}
  \item \url{https://tjuheide.github.io/MCC/}
  \item Report bugs at \url{https://github.com/tjuheide/MCC/issues}
}

}
\author{
\strong{Maintainer}: Uffe Heide-Jørgensen \email{uhj@clin.au.dk} (\href{https://orcid.org/0000-0002-8407-4618}{ORCID})

}
\keyword{internal}
