% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mcc.R
\name{mcc}
\alias{mcc}
\title{Modal Clustering for Categorical Data}
\usage{
mcc(data, min_overlap = 1, rounding = 2, independence_tol = 1e-04)
}
\arguments{
\item{data}{A data frame containing only categorical variables.}

\item{min_overlap}{Minimum number of shared categories required to connect two cells (default = 1).}

\item{rounding}{Number of decimal places used when checking for independence (default = 2).}

\item{independence_tol}{Tolerance level for determining independence between cells (default = 0.0001).}
}
\value{
A list with the following components:
\describe{
\item{original_data}{Original data frame.}
\item{individual_data_with_cl}{Individual-level data with assigned cluster labels.}
\item{frequency_data_with_cl}{Frequency-level data with cluster assignments.}
\item{edge_list}{Data frame of edges with weights used to build the graph.}
\item{graph}{igraph object representing the connectivity among categorical combinations.}
\item{clustering}{Final clustering labels (if mode-shift and level-set results coincide).}
}
}
\description{
Performs modal clustering (mode-shift and level-set) on categorical data
as introduced by Corsini and Menardi (2025).
}
\details{
This function constructs a weighted directed graph whose nodes represent unique
categorical combinations and whose edges connect statistically dependent combinations.
The direction of each edge is determined by relative density. Clustering is then performed
using both mode-shift and level-set procedures. If both yield the same partition, that
is returned as the final clustering.
}
\examples{
\dontrun{
data <- data.frame(A = sample(letters[1:3], 100, TRUE),
                   B = sample(LETTERS[1:2], 100, TRUE))
res <- mcc(data)
}

}
\references{
Corsini, N. and Menardi, G. (2025). \emph{Modal Clustering for Categorical Data.}
}
