% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Data.R
\docType{data}
\name{mcc_data}
\alias{mcc_data}
\title{Sample Recurrent Event Data}
\format{
A data.frame containing three fields:
\describe{
  \item{idx}{Unique subject identifier.}
  \item{time}{Observation time between 0 and 10.}
  \item{arm}{Treatment arm, 0 for reference, 1 for treatment.}
  \item{status}{Status indicator, 0 for censoring, 1 for an event, 2 for death.}
}
}
\usage{
data(mcc_data)
}
\description{
Sample recurrent events data. Events occurred in the treatment arm according
to a Poisson process with rate 1.00, and in the reference arm according to a
Poisson process with rate 1.35. Data were truncated at time 10. Death and
censoring occurred independently according to an exponential distribution
with rate 0.25 in each arm. The number of subjects in each arm is 100.
}
\keyword{datasets}
