% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/PlotMCF.R
\name{MCFCurve}
\alias{MCFCurve}
\title{MCF Curve}
\usage{
MCFCurve(data, idx_name = "idx", status_name = "status", time_name = "time")
}
\arguments{
\item{data}{Data.frame.}

\item{idx_name}{Name of index (subject identifier) column in data.}

\item{status_name}{Name of status column in data.}

\item{time_name}{Name of column column in data.}
}
\value{
stepfun.
}
\description{
Construct a function to evaluate the mean cumulative function at a given time
for a single treatment arm.
}
