% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/PlotMCF.R
\name{MCFPlotFrame}
\alias{MCFPlotFrame}
\title{MCF Plotting Frame}
\usage{
MCFPlotFrame(
  data,
  eval_points = 1000,
  idx_name = "idx",
  status_name = "status",
  tau = NULL,
  time_name = "time"
)
}
\arguments{
\item{data}{Data.frame.}

\item{eval_points}{Number of points at which to evaluate the curve.}

\item{idx_name}{Name of index (subject identifier) column in data.}

\item{status_name}{Name of status column in data.}

\item{tau}{Truncation time.}

\item{time_name}{Name of column column in data.}
}
\value{
Data.frame containing `time` and `mcf`.
}
\description{
Constructs the MCF plotting frame for a single treatment arm.
}
