% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Bootstrap.R
\name{PermData}
\alias{PermData}
\title{Permute Treatment Assignments.}
\usage{
PermData(data)
}
\arguments{
\item{data}{Data.frame.}
}
\value{
Bootstrapped data.frame.
}
\description{
Permute Treatment Assignments.
}
