% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Resampling.R
\name{PermData}
\alias{PermData}
\title{Permute Treatment Assignments.}
\usage{
PermData(data)
}
\arguments{
\item{data}{Data.frame containing 'arm' and 'idx'.}
}
\value{
Bootstrapped data.frame.
}
\description{
Permute Treatment Assignments.
}
