% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/PlotMCF2.R
\name{PlotAUMCFs}
\alias{PlotAUMCFs}
\title{Plot Area Under the Mean Cumulative Function}
\usage{
PlotAUMCFs(
  data,
  which_arm,
  arm_label = "Placebo",
  arm_name = "arm",
  color = "#C65842",
  idx_name = "idx",
  status_name = "status",
  strata_name = NULL,
  time_name = "time",
  title = NULL,
  tau = NULL,
  weights = NULL,
  x_breaks = NULL,
  x_lim = NULL,
  x_name = "Time",
  y_breaks = NULL,
  y_lim = NULL,
  y_name = "Mean Cumulative Count"
)
}
\arguments{
\item{data}{Data.frame.}

\item{which_arm}{Arm to plot.}

\item{arm_label}{Label for the arm.}

\item{arm_name}{Name of arm column in data.}

\item{color}{Color.}

\item{idx_name}{Name of index (subject identifier) column in data.}

\item{status_name}{Name of status column in data.}

\item{strata_name}{Name of stratum column in data.}

\item{time_name}{Name of column column in data.}

\item{title}{Plot title.}

\item{tau}{Truncation time for shading.}

\item{weights}{Optional column of weights, controlling the size of the jump
in the cumulative count curve at times with status == 1.}

\item{x_breaks}{X-axis breaks.}

\item{x_lim}{X-axis limits.}

\item{x_name}{X-axis label.}

\item{y_breaks}{Y-axis breaks.}

\item{y_lim}{Y-axis limits.}

\item{y_name}{Y-axis label.}
}
\value{
ggplot object.
}
\description{
Plot area under the mean cumulative function for a single treatment arm.
}
