% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/PlotMCF.R
\name{PlotAUMCFs}
\alias{PlotAUMCFs}
\title{Plot AUMCFs.}
\usage{
PlotAUMCFs(
  data,
  which_arm,
  arm_label = "Placebo",
  arm_name = "arm",
  color = "#C65842",
  idx_name = "idx",
  status_name = "status",
  time_name = "time",
  title = NULL,
  tau = NULL,
  x_breaks = NULL,
  x_max = NULL,
  x_name = "Time",
  y_lim = NULL,
  y_name = "Mean Cumulative Count"
)
}
\arguments{
\item{data}{Data.frame.}

\item{which_arm}{Arm to plot.}

\item{arm_label}{Label for the arm.}

\item{arm_name}{Name of arm column in data.}

\item{color}{Color.}

\item{idx_name}{Name of index (subject identifier) column in data.}

\item{status_name}{Name of status column in data.}

\item{time_name}{Name of column column in data.}

\item{title}{Plot title.}

\item{tau}{Truncation time for shading.}

\item{x_breaks}{X-axis breaks.}

\item{x_max}{X-axis upper limit.}

\item{x_name}{X-axis label.}

\item{y_lim}{Y-axis limits.}

\item{y_name}{Y-axis label.}
}
\value{
ggplot object.
}
\description{
Plot area under the mean cumulative function for a single treatment arm.
}
