% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/PlotMCF.R
\name{PlotMCFs}
\alias{PlotMCFs}
\title{Plot MCFs.}
\usage{
PlotMCFs(
  data,
  arm_name = "arm",
  color_labs = c("Ctrl", "Trt"),
  ctrl_color = "#C65842",
  idx_name = "idx",
  status_name = "status",
  strata_name = NULL,
  tau = NULL,
  time_name = "time",
  title = NULL,
  trt_color = "#6385B8",
  x_breaks = NULL,
  x_lim = NULL,
  x_name = "Time",
  y_breaks = NULL,
  y_lim = NULL,
  y_name = "Mean Cumulative Count"
)
}
\arguments{
\item{data}{Data.frame.}

\item{arm_name}{Name of arm column in data.}

\item{color_labs}{Color labels.}

\item{ctrl_color}{Color for control arm.}

\item{idx_name}{Name of index (subject identifier) column in data.}

\item{status_name}{Name of status column in data.}

\item{strata_name}{Name of stratum column in data.}

\item{tau}{Truncation time.}

\item{time_name}{Name of column column in data.}

\item{title}{Plot title.}

\item{trt_color}{Color for treatment arm.}

\item{x_breaks}{X-axis breaks.}

\item{x_lim}{X-axis limits.}

\item{x_name}{X-axis label.}

\item{y_breaks}{Y-axis breaks.}

\item{y_lim}{Y-axis limits.}

\item{y_name}{Y-axis label.}
}
\value{
ggplot object.
}
\description{
Plot the mean cumulative functions comparing two treatment arms.
}
