% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/PlotMCF1.R
\name{PlotOneSampleAUMCF}
\alias{PlotOneSampleAUMCF}
\title{Plot One Sample Area Under the Mean Cumulative Function}
\usage{
PlotOneSampleAUMCF(
  data,
  color = "#C65842",
  color_lab = "Arm",
  idx_name = "idx",
  status_name = "status",
  time_name = "time",
  title = NULL,
  tau = NULL,
  weights = NULL,
  x_breaks = NULL,
  x_lim = NULL,
  x_name = "Time",
  y_breaks = NULL,
  y_lim = NULL,
  y_name = "Mean Cumulative Count"
)
}
\arguments{
\item{data}{Data.frame.}

\item{color}{Color.}

\item{color_lab}{Color label.}

\item{idx_name}{Name of index (subject identifier) column in data.}

\item{status_name}{Name of status column in data.}

\item{time_name}{Name of column column in data.}

\item{title}{Plot title.}

\item{tau}{Truncation time for shading.}

\item{weights}{Optional column of weights, controlling the size of the jump
in the cumulative count curve at times with status == 1.}

\item{x_breaks}{X-axis breaks.}

\item{x_lim}{X-axis limits.}

\item{x_name}{X-axis label.}

\item{y_breaks}{Y-axis breaks.}

\item{y_lim}{Y-axis limits.}

\item{y_name}{Y-axis label.}
}
\value{
ggplot object.
}
\description{
Plot area under the mean cumulative function for a single treatment arm.
}
