% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Mean-cumulative-function.R
\name{TabulateEvents}
\alias{TabulateEvents}
\title{Tabulate Events}
\usage{
TabulateEvents(time, status, idx)
}
\arguments{
\item{time}{Event time.}

\item{status}{Status, coded as 1 for the recurrent event, 0 otherwise.}

\item{idx}{Subject index.}
}
\value{
Data.frame with these columns:
\itemize{
  \item `event_times` distinct recurrent event times.
  \item `nar` number at risk at the event time. 
  \item `events` number of events that occurred. 
}
}
\description{
Tabulate the number at risk and the number of events at each unique
observed event time.
}
